fix: make the return type of scan function nullable
Issue
=====

Looking at the return type of `scan()`, it is not nullable, which hints
at it throwing an error. However, in practice it does not throw an error
and returns `null` when the path does not exist.

Solution
========

Make the return type of `scan()` nullable to indicate that the caller
should check if it succeeded.